refactor: swap screen
